# solution_example/account-management-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for managing accounts, reservations, products and operations.",
        "title": "Account Management Service",
        "contact": {
            "name": "Changaz Danial",
            "email": "changaz.d@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8089",
    "basePath": "/",
    "paths": {
...
        "/api/v1/reservations/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Create reservation",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account-management-service_internal_controller_http_v1.reservationCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/account-management-service_internal_controller_http_v1.reservationRoutes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/reservations/refund": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Refund reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Refund reservation",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account-management-service_internal_controller_http_v1.reservationRefundInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/reservations/revenue": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Revenue reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Revenue reservation",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account-management-service_internal_controller_http_v1.reservationRevenueInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "account-management-service_internal_controller_http_v1.reservationCreateInput": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "order_id",
                "product_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "account-management-service_internal_controller_http_v1.reservationRefundInput": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "account-management-service_internal_controller_http_v1.reservationRevenueInput": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "order_id",
                "product_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "account-management-service_internal_controller_http_v1.reservationRoutes": {
            "type": "object"
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}